LAB2/source/requirements.txt
# Core dependencies
numpy>=1.21.0
pillow>=8.0.0
opencv-python>=4.5.0
scikit-image>=0.18.0
scipy>=1.7.0

# Deep Learning
# PyTorch - uncomment one of the following based on your system
# CPU only:
# torch>=1.12.0
# torchvision>=0.13.0
# torchaudio>=0.12.0

# CUDA 11.8:
# torch>=2.0.0+cu118 --index-url https://download.pytorch.org/whl/cu118
# torchvision>=0.15.0+cu118 --index-url https://download.pytorch.org/whl/cu118
# torchaudio>=2.0.0+cu118 --index-url https://download.pytorch.org/whl/cu118

# CUDA 12.1:
# torch>=2.0.0+cu121 --index-url https://download.pytorch.org/whl/cu121
# torchvision>=0.15.0+cu121 --index-url https://download.pytorch.org/whl/cu121
# torchaudio>=2.0.0+cu121 --index-url https://download.pytorch.org/whl/cu121

# Utilities
tqdm>=4.62.0
matplotlib>=3.4.0
pandas>=1.3.0

